Conflicts:
	Kernel/kernel.elf
	Userland/0000-userModule.elf
	Userland/UserModule/standard_library/commands.c

+++ Userland/UserModule/standard_library/mystdlib.c
@@ -125,7 +125,7 @@ long strToNumPos(char* string){
     return num;
 }
 
-long strToNum(char* string){
+long long strToNum(char* string){
     int isNegative=0;
     long num = 0;
     size_t i=0, length = strlen(string);
